refactor(types): extract V1/V2 namespaces to isolated files
- V1 extracted to v1/index.ts with own BaseStruct/BaseTypes
- V2 extracted to v2/index.ts with own BaseStruct/BaseTypes
- V1 kept only original types (Struct<T,U>, ArrayStruct<T,U>, etc.)
- V3 convenience aliases removed from V1 (use V3.* directly)
- types/index.ts imports V1/V2 via `import type * as`, re-exports them
- Added top-level GenericStruct/ObjectStruct/Struct/StructType aliases to V3
  for backward compat with consumers importing bare names
